pyproject.toml
[project]
name = "genesis-ros"
version = "0.1.0"
description = "ROS2 bridge / related tools for genesis"
license = "Apache-2.0"
readme = "README.md"
requires-python = "==3.10.12"
dependencies = [
    "amber-mcap>=0.6.0; sys_platform == 'linux'",
    "torch==2.5.0+cu124; sys_platform == 'linux'",
    "torch==2.5.0; sys_platform == 'darwin'",
    "torchaudio==2.5.0+cu124; sys_platform == 'linux'",
    "torchaudio==2.5.0; sys_platform == 'darwin'",
    "torchvision==0.20.0+cu124; sys_platform == 'linux'",
    "torchvision==0.20.0; sys_platform == 'darwin'",
    "xacrodoc>=0.4.1",
    "genesis-world>=0.2.0",
    "rsl-rl-lib==2.2.4",
    "trimesh>=4.6.8",
    "dataclass-wizard>=0.22.3",
    "taichi>=1.7.3",
    "eclipse-zenoh>=1.3.4",
    "pycdr2>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
preprocess_urdf = "genesis_ros.preprocess_urdf:main"
simulation = "genesis_ros.run_simulation:main"
ppo_train = "genesis_ros.ppo.ppo_train:cli_entrypoint"
ppo_eval = "genesis_ros.ppo.ppo_eval:cli_entrypoint"

[tool.uv]
find-links = [
    "https://download.pytorch.org/whl/cu124/torch",
    "https://download.pytorch.org/whl/cu124/torchvision",
    "https://download.pytorch.org/whl/cu124/torchaudio",
]
default-groups = ["dev"]

[tool.coverage.html]
directory = "_htmlcov"

[tool.coverage.xml]
output = "_htmlcov/coverage.xml"

[tool.coverage.report]
exclude_lines = [
    "@overload",
    "if TYPE_CHECKING:",
    "pragma: no cover",
    "raise AssertionError()",
    "return NotImplemented",
]

[tool.pytest.ini_options]
addopts = [
    "--cov-report=html",
    "--cov-report=xml",
    "--cov-report=term",
    "--cov=genesis_ros",
    "--doctest-modules",
    "--import-mode=importlib",
]
testpaths = ["tests",]

[dependency-groups]
dev = [
    "black==25.1.0",
    "mypy==1.15.0",
    "pytest==8.3.5",
    "pre-commit>=4.2.0",
    "lark>=1.2.2",
    "pytest-cov>=6.1.1",
    "pytest-testmon>=2.1.3",
]
